Use add_dummy _variable for ASCAT (#42)
### This PR updates the following:

**Python script for ASCAT:**
- Use add_dummy_variable and map_path
- Consolidate the addition of new variables in one function
(_add_wind_obs)
- Use a dictionary to add dummy variables

**YAML for ASCAT:**
- Make sure the data type of wind directory is float
- Update global attributes (to follow IODA convention)

**Python script for SSMIS:**
- Use a dictionary to add dummy variables

**Python script for adpupa_prepbufr**
- Fix coding norms

```
            dummy_mappings = [
               ('solarZenithAngle', 'latitude'),
               ('solarAzimuthAngle', 'latitude'),
               ('sensorZenithAngle', 'latitude'),
               ('sensorAzimuthAngle','latitude')
            ]
            for target_var, source_var in dummy_mappings:
                add_dummy_variable(container, target_var, category, source_var)

```


Do not change results.

---------
